Moved more methods from util.Util
