+++ src/main/java/net/sf/jabref/gui/ContentSelectorDialog2.java
@@ -46,6 +46,7 @@ import net.sf.jabref.MetaData;
 import net.sf.jabref.gui.help.HelpFiles;
 import net.sf.jabref.gui.help.HelpAction;
 import net.sf.jabref.gui.util.FocusRequester;
+import net.sf.jabref.gui.util.KeyBinder;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.util.Util;
 
@@ -114,7 +115,7 @@ class ContentSelectorDialog2 extends JDialog {
         setupFieldSelector();
         setupWordSelector();
         setupActions();
-        Util.bindCloseDialogKeyToCancelAction(this.rootPane, cancel.getAction());
+        KeyBinder.bindCloseDialogKeyToCancelAction(this.rootPane, cancel.getAction());
         int fieldInd = fieldListModel.indexOf(currentField);
         if (fieldInd >= 0) {
             fieldList.setSelectedIndex(fieldInd);
