+++ src/main/java/net/sf/jabref/gui/preftabs/ImportSettingsTab.java
@@ -19,6 +19,7 @@ import java.awt.BorderLayout;
 import javax.swing.*;
 
 import net.sf.jabref.Globals;
+import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.pdfimport.ImportDialog;
 
@@ -27,11 +28,8 @@ import com.jgoodies.forms.layout.FormLayout;
 
 public class ImportSettingsTab extends JPanel implements PrefsTab {
 
-    public static final String PREF_IMPORT_ALWAYSUSE = "importAlwaysUsePDFImportStyle";
-    public static final String PREF_IMPORT_DEFAULT_PDF_IMPORT_STYLE = "importDefaultPDFimportStyle";
     public static final int DEFAULT_STYLE = ImportDialog.CONTENT;
 
-    public static final String PREF_IMPORT_FILENAMEPATTERN = "importFileNamePattern";
     private static final String[] DEFAULT_FILENAMEPATTERNS_DISPLAY = new String[] {
             "bibtexkey",
             "bibtexkey - title",
@@ -108,8 +106,8 @@ public class ImportSettingsTab extends JPanel implements PrefsTab {
 
     @Override
     public void setValues() {
-        useDefaultPDFImportStyle.setSelected(Globals.prefs.getBoolean(ImportSettingsTab.PREF_IMPORT_ALWAYSUSE));
-        int style = Globals.prefs.getInt(ImportSettingsTab.PREF_IMPORT_DEFAULT_PDF_IMPORT_STYLE);
+        useDefaultPDFImportStyle.setSelected(Globals.prefs.getBoolean(JabRefPreferences.PREF_IMPORT_ALWAYSUSE));
+        int style = Globals.prefs.getInt(JabRefPreferences.PREF_IMPORT_DEFAULT_PDF_IMPORT_STYLE);
         switch (style) {
         case ImportDialog.NOMETA:
             radioButtonNoMeta.setSelected(true);
@@ -128,12 +126,12 @@ public class ImportSettingsTab extends JPanel implements PrefsTab {
             radioButtonPDFcontent.setSelected(true);
             break;
         }
-        fileNamePattern.setText(Globals.prefs.get(ImportSettingsTab.PREF_IMPORT_FILENAMEPATTERN));
+        fileNamePattern.setText(Globals.prefs.get(JabRefPreferences.PREF_IMPORT_FILENAMEPATTERN));
     }
 
     @Override
     public void storeSettings() {
-        Globals.prefs.putBoolean(ImportSettingsTab.PREF_IMPORT_ALWAYSUSE, useDefaultPDFImportStyle.isSelected());
+        Globals.prefs.putBoolean(JabRefPreferences.PREF_IMPORT_ALWAYSUSE, useDefaultPDFImportStyle.isSelected());
         int style = ImportSettingsTab.DEFAULT_STYLE;
         if (radioButtonNoMeta.isSelected()) {
             style = ImportDialog.NOMETA;
@@ -144,8 +142,8 @@ public class ImportSettingsTab extends JPanel implements PrefsTab {
         } else if (radioButtononlyAttachPDF.isSelected()) {
             style = ImportDialog.ONLYATTACH;
         }
-        Globals.prefs.putInt(ImportSettingsTab.PREF_IMPORT_DEFAULT_PDF_IMPORT_STYLE, style);
-        Globals.prefs.put(ImportSettingsTab.PREF_IMPORT_FILENAMEPATTERN, fileNamePattern.getText());
+        Globals.prefs.putInt(JabRefPreferences.PREF_IMPORT_DEFAULT_PDF_IMPORT_STYLE, style);
+        Globals.prefs.put(JabRefPreferences.PREF_IMPORT_FILENAMEPATTERN, fileNamePattern.getText());
     }
 
     @Override
