+++ src/main/java/net/sf/jabref/logic/util/io/FileUtil.java
@@ -18,11 +18,17 @@ package net.sf.jabref.logic.util.io;
 import net.sf.jabref.BibDatabaseContext;
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
+import net.sf.jabref.logic.journals.JournalAbbreviationRepository;
+import net.sf.jabref.logic.layout.Layout;
+import net.sf.jabref.logic.layout.LayoutHelper;
 import net.sf.jabref.logic.util.OS;
+import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.model.entry.BibEntry;
 
 import net.sf.jabref.model.entry.FileField;
 import net.sf.jabref.model.entry.ParsedFileField;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
@@ -30,10 +36,14 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.StringReader;
 import java.util.*;
 import java.util.regex.Pattern;
 
 public class FileUtil {
+
+    private static final Log LOGGER = LogFactory.getLog(FileUtil.class);
+
     private static final String FILE_SEPARATOR = System.getProperty("file.separator");
     private static final Pattern SLASH = Pattern.compile("/");
     private static final Pattern BACKSLASH = Pattern.compile("\\\\");
@@ -358,6 +368,32 @@ public class FileUtil {
         }
 
         return result;
+    }
+
+    /**
+     * Determines filename provided by an entry in a database
+     *
+     * @param database the database, where the entry is located
+     * @param entry    the entry to which the file should be linked to
+     * @param repository
+     * @return a suggested fileName
+     */
+    public static String getLinkedFileName(BibDatabase database, BibEntry entry,
+            JournalAbbreviationRepository repository) {
+        String targetName = entry.getCiteKey() == null ? "default" : entry.getCiteKey();
+        StringReader sr = new StringReader(Globals.prefs.get(JabRefPreferences.PREF_IMPORT_FILENAMEPATTERN));
+        Layout layout = null;
+        try {
+            layout = new LayoutHelper(sr, repository).getLayoutFromText();
+        } catch (IOException e) {
+            LOGGER.info("Wrong format " + e.getMessage(), e);
+        }
+        if (layout != null) {
+            targetName = layout.doLayout(entry, database);
+        }
+        //Removes illegal characters from filename
+        targetName = FileNameCleaner.cleanFileName(targetName);
+        return targetName;
     }
 
 }
